+++ src/types/QuickPickItem.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode';
 export interface TodoQuickPickItem extends vscode.QuickPickItem {
     line: number;
+    filePath?: string;
 }